uupdating

+++ source/Backend/src/main/java/com/acare/backend/repository/ServiceRepository.java
@@ -1,5 +1,6 @@
 package com.acare.backend.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -10,5 +11,5 @@ public interface ServiceRepository extends JpaRepository<Service, Long> {
     List<Service> findByName(String name);
     List<Service> findAllByNameContainingIgnoreCase(String name);
     List<Service> findByActiveTrue();
-    List<Service> findByPriceBetween(Double min, Double max);
+    List<Service> findByPriceBetween(BigDecimal min, BigDecimal max);
 }
